Split data_file.go into reader.go and writer.go
